Разбиение на компоненты
