+++ project/src/components/footer/footer.tsx
@@ -1,0 +1,15 @@
+import Logo from '../../components/logo/logo';
+
+function Footer() {
+  return (
+    <footer className="page-footer">
+      <Logo isLight />
+
+      <div className="copyright">
+        <p>© 2019 What to watch Ltd.</p>
+      </div>
+    </footer>
+  );
+}
+
+export default Footer;
