+++ project/src/pages/film/film.tsx
@@ -6,15 +6,19 @@ import {Link, useLocation} from 'react-router-dom';
 import Tabs from '../../components/tabs/tabs';
 import User from '../../components/user/user';
 import { useAppSelector } from '../../hooks';
+import Footer from '../../components/footer/footer';
 
 type typeProps = {
   filmId?: number,
 }
 
 function Film({filmId = 0}: typeProps) {
+  /* eslint-disable no-console */
   const MAX_FILMS = 4;
   const {films} = useAppSelector((state) => state);
+  console.log(films);
   const {backgroundColor, backgroundImage, name, genre, released, posterImage} = films[filmId];
+  console.log(films);
 
   let location = useLocation().hash.substr(1);
   location = TabNames.Overview;
@@ -109,13 +113,7 @@ function Film({filmId = 0}: typeProps) {
           <CardList films={films.slice(0, MAX_FILMS)} />
         </section>
 
-        <footer className="page-footer">
-          <Logo isLight />
-
-          <div className="copyright">
-            <p>© 2019 What to watch Ltd.</p>
-          </div>
-        </footer>
+        <Footer />
       </div>
     </>
   );
