+++ project/src/pages/my-list/my-list.tsx
@@ -1,4 +1,5 @@
 import CardList from '../../components/card-list/card-list';
+import Footer from '../../components/footer/footer';
 import Logo from '../../components/logo/logo';
 import User from '../../components/user/user';
 import { useAppSelector } from '../../hooks';
@@ -24,13 +25,7 @@ function MyList() {
         </div>
       </section>
 
-      <footer className="page-footer">
-        <Logo isLight />
-
-        <div className="copyright">
-          <p>© 2019 What to watch Ltd.</p>
-        </div>
-      </footer>
+      <Footer />
     </div>
   );
 }
